[KCP] reorganize unit tests
