+++ controlplane/kubeadm/controllers/status_test.go
@@ -1,0 +1,273 @@
+/*
+Copyright 2020 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package controllers
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	. "github.com/onsi/gomega"
+
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/tools/record"
+	"k8s.io/klog/klogr"
+	clusterv1 "sigs.k8s.io/cluster-api/api/v1alpha3"
+	controlplanev1 "sigs.k8s.io/cluster-api/controlplane/kubeadm/api/v1alpha3"
+	"sigs.k8s.io/cluster-api/controlplane/kubeadm/internal"
+	"sigs.k8s.io/controller-runtime/pkg/log"
+)
+
+func TestKubeadmControlPlaneReconciler_updateStatusNoMachines(t *testing.T) {
+	g := NewWithT(t)
+
+	cluster := &clusterv1.Cluster{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "foo",
+			Namespace: "test",
+		},
+	}
+
+	kcp := &controlplanev1.KubeadmControlPlane{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: cluster.Namespace,
+			Name:      "foo",
+		},
+		Spec: controlplanev1.KubeadmControlPlaneSpec{
+			Version: "v1.16.6",
+		},
+	}
+	kcp.Default()
+	g.Expect(kcp.ValidateCreate()).To(Succeed())
+
+	fakeClient := newFakeClient(g, kcp.DeepCopy(), cluster.DeepCopy())
+	log.SetLogger(klogr.New())
+
+	r := &KubeadmControlPlaneReconciler{
+		Client: fakeClient,
+		Log:    log.Log,
+		scheme: scheme.Scheme,
+		managementCluster: &fakeManagementCluster{
+			Machines: map[string]*clusterv1.Machine{},
+			Workload: fakeWorkloadCluster{},
+		},
+		recorder: record.NewFakeRecorder(32),
+	}
+
+	g.Expect(r.updateStatus(context.Background(), kcp, cluster)).To(Succeed())
+	g.Expect(kcp.Status.Replicas).To(BeEquivalentTo(0))
+	g.Expect(kcp.Status.ReadyReplicas).To(BeEquivalentTo(0))
+	g.Expect(kcp.Status.UnavailableReplicas).To(BeEquivalentTo(0))
+	g.Expect(kcp.Status.Initialized).To(BeFalse())
+	g.Expect(kcp.Status.Ready).To(BeFalse())
+	g.Expect(kcp.Status.Selector).NotTo(BeEmpty())
+	g.Expect(kcp.Status.FailureMessage).To(BeNil())
+	g.Expect(kcp.Status.FailureReason).To(BeEquivalentTo(""))
+}
+
+func TestKubeadmControlPlaneReconciler_updateStatusAllMachinesNotReady(t *testing.T) {
+	g := NewWithT(t)
+
+	cluster := &clusterv1.Cluster{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "foo",
+			Namespace: "test",
+		},
+	}
+
+	kcp := &controlplanev1.KubeadmControlPlane{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: cluster.Namespace,
+			Name:      "foo",
+		},
+		Spec: controlplanev1.KubeadmControlPlaneSpec{
+			Version: "v1.16.6",
+		},
+	}
+	kcp.Default()
+	g.Expect(kcp.ValidateCreate()).To(Succeed())
+
+	machines := map[string]*clusterv1.Machine{}
+	objs := []runtime.Object{cluster.DeepCopy(), kcp.DeepCopy()}
+	for i := 0; i < 3; i++ {
+		name := fmt.Sprintf("test-%d", i)
+		m, n := createMachineNodePair(name, cluster, kcp, false)
+		objs = append(objs, n)
+		machines[m.Name] = m
+	}
+
+	fakeClient := newFakeClient(g, objs...)
+	log.SetLogger(klogr.New())
+
+	r := &KubeadmControlPlaneReconciler{
+		Client: fakeClient,
+		Log:    log.Log,
+		scheme: scheme.Scheme,
+		managementCluster: &fakeManagementCluster{
+			Machines: machines,
+			Workload: fakeWorkloadCluster{},
+		},
+		recorder: record.NewFakeRecorder(32),
+	}
+
+	g.Expect(r.updateStatus(context.Background(), kcp, cluster)).To(Succeed())
+	g.Expect(kcp.Status.Replicas).To(BeEquivalentTo(3))
+	g.Expect(kcp.Status.ReadyReplicas).To(BeEquivalentTo(0))
+	g.Expect(kcp.Status.UnavailableReplicas).To(BeEquivalentTo(3))
+	g.Expect(kcp.Status.Selector).NotTo(BeEmpty())
+	g.Expect(kcp.Status.FailureMessage).To(BeNil())
+	g.Expect(kcp.Status.FailureReason).To(BeEquivalentTo(""))
+	g.Expect(kcp.Status.Initialized).To(BeFalse())
+	g.Expect(kcp.Status.Ready).To(BeFalse())
+}
+
+func TestKubeadmControlPlaneReconciler_updateStatusAllMachinesReady(t *testing.T) {
+	g := NewWithT(t)
+
+	cluster := &clusterv1.Cluster{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: "test",
+			Name:      "foo",
+		},
+	}
+
+	kcp := &controlplanev1.KubeadmControlPlane{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: cluster.Namespace,
+			Name:      "foo",
+		},
+		Spec: controlplanev1.KubeadmControlPlaneSpec{
+			Version: "v1.16.6",
+		},
+	}
+	kcp.Default()
+	g.Expect(kcp.ValidateCreate()).To(Succeed())
+
+	objs := []runtime.Object{cluster.DeepCopy(), kcp.DeepCopy(), kubeadmConfigMap()}
+	machines := map[string]*clusterv1.Machine{}
+	for i := 0; i < 3; i++ {
+		name := fmt.Sprintf("test-%d", i)
+		m, n := createMachineNodePair(name, cluster, kcp, true)
+		objs = append(objs, n)
+		machines[m.Name] = m
+	}
+
+	fakeClient := newFakeClient(g, objs...)
+	log.SetLogger(klogr.New())
+
+	r := &KubeadmControlPlaneReconciler{
+		Client: fakeClient,
+		Log:    log.Log,
+		scheme: scheme.Scheme,
+		managementCluster: &fakeManagementCluster{
+			Machines: machines,
+			Workload: fakeWorkloadCluster{
+				Status: internal.ClusterStatus{
+					Nodes:            3,
+					ReadyNodes:       3,
+					HasKubeadmConfig: true,
+				},
+			},
+		},
+		recorder: record.NewFakeRecorder(32),
+	}
+
+	g.Expect(r.updateStatus(context.Background(), kcp, cluster)).To(Succeed())
+	g.Expect(kcp.Status.Replicas).To(BeEquivalentTo(3))
+	g.Expect(kcp.Status.ReadyReplicas).To(BeEquivalentTo(3))
+	g.Expect(kcp.Status.UnavailableReplicas).To(BeEquivalentTo(0))
+	g.Expect(kcp.Status.Selector).NotTo(BeEmpty())
+	g.Expect(kcp.Status.FailureMessage).To(BeNil())
+	g.Expect(kcp.Status.FailureReason).To(BeEquivalentTo(""))
+	g.Expect(kcp.Status.Initialized).To(BeTrue())
+	g.Expect(kcp.Status.Ready).To(BeTrue())
+}
+
+func TestKubeadmControlPlaneReconciler_updateStatusMachinesReadyMixed(t *testing.T) {
+	g := NewWithT(t)
+
+	cluster := &clusterv1.Cluster{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "foo",
+			Namespace: "test",
+		},
+	}
+
+	kcp := &controlplanev1.KubeadmControlPlane{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: cluster.Namespace,
+			Name:      "foo",
+		},
+		Spec: controlplanev1.KubeadmControlPlaneSpec{
+			Version: "v1.16.6",
+		},
+	}
+	kcp.Default()
+	g.Expect(kcp.ValidateCreate()).To(Succeed())
+	machines := map[string]*clusterv1.Machine{}
+	objs := []runtime.Object{cluster.DeepCopy(), kcp.DeepCopy()}
+	for i := 0; i < 4; i++ {
+		name := fmt.Sprintf("test-%d", i)
+		m, n := createMachineNodePair(name, cluster, kcp, false)
+		machines[m.Name] = m
+		objs = append(objs, n)
+	}
+	m, n := createMachineNodePair("testReady", cluster, kcp, true)
+	objs = append(objs, n, kubeadmConfigMap())
+	machines[m.Name] = m
+	fakeClient := newFakeClient(g, objs...)
+	log.SetLogger(klogr.New())
+
+	r := &KubeadmControlPlaneReconciler{
+		Client: fakeClient,
+		Log:    log.Log,
+		scheme: scheme.Scheme,
+		managementCluster: &fakeManagementCluster{
+			Machines: machines,
+			Workload: fakeWorkloadCluster{
+				Status: internal.ClusterStatus{
+					Nodes:            5,
+					ReadyNodes:       1,
+					HasKubeadmConfig: true,
+				},
+			},
+		},
+		recorder: record.NewFakeRecorder(32),
+	}
+
+	g.Expect(r.updateStatus(context.Background(), kcp, cluster)).To(Succeed())
+	g.Expect(kcp.Status.Replicas).To(BeEquivalentTo(5))
+	g.Expect(kcp.Status.ReadyReplicas).To(BeEquivalentTo(1))
+	g.Expect(kcp.Status.UnavailableReplicas).To(BeEquivalentTo(4))
+	g.Expect(kcp.Status.Selector).NotTo(BeEmpty())
+	g.Expect(kcp.Status.FailureMessage).To(BeNil())
+	g.Expect(kcp.Status.FailureReason).To(BeEquivalentTo(""))
+	g.Expect(kcp.Status.Initialized).To(BeTrue())
+	g.Expect(kcp.Status.Ready).To(BeTrue())
+}
+
+func kubeadmConfigMap() *corev1.ConfigMap {
+	return &corev1.ConfigMap{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "kubeadm-config",
+			Namespace: metav1.NamespaceSystem,
+		},
+	}
+}
